Revert "Dynamic import of contactFormValidation & quoteFormValidation which triggers loading of zod"
This reverts commit c759f0324a8b4bdfa9164c5002b666f939f3cb74.
