+++ composables/useQuoteForm.ts
@@ -5,16 +5,107 @@
  *
  * Provides form state management, validation, and submission handling
  * for the quote request form with Zod schema validation.
- * 
- * Note: Zod validation is dynamically imported to keep it out of the critical bundle.
  */
 
+import { z } from 'zod'
 import { useQuoteFormStore } from '~/stores/quoteFormStore'
-import type { QuoteFormData, FormState } from '~/types/Forms'
 
 // ===== CONSTANTS =====
+/** Maximum file size: 7MB */
+const MAX_FILE_SIZE = 7 * 1024 * 1024
+
 /** Maximum number of images allowed */
 export const MAX_IMAGE_COUNT = 16
+
+/** Allowed image MIME types */
+const ALLOWED_IMAGE_TYPES = ['image/jpeg', 'image/jpg', 'image/png', 'image/webp', 'image/gif']
+
+// ===== ZOD SCHEMA =====
+/** Quote form validation schema */
+export const quoteFormSchema = z.object({
+  name: z
+    .string()
+    .min(2, 'Namnet måste vara minst 2 tecken')
+    .max(100, 'Namnet får inte vara längre än 100 tecken'),
+  email: z
+    .string()
+    .email('Ange en giltig e-postadress')
+    .min(1, 'E-post är obligatoriskt'),
+  phone: z
+    .string()
+    .regex(/^(\+\d{1,3})?[\s]*-?[\s]*(\(?\d{1,4}\)?)?[\s]*-?[\s]*\d[\d\s-]{5,14}$/, 'Ange ett giltigt telefonnummer')
+    .optional()
+    .or(z.literal('')),
+  customerType: z
+    .enum(['Privatperson', 'Företag'], {
+      errorMap: () => ({ message: 'Välj om du är privatperson eller företag' }),
+    }),
+  subject: z.literal('Offertförfrågan'),
+  productCategory: z
+    .string()
+    .optional()
+    .or(z.literal('')),
+  product: z
+    .string()
+    .optional()
+    .or(z.literal('')),
+  productId: z
+    .string()
+    .regex(/^[A-Za-z0-9_-]+$/, 'Produkt ID får endast innehålla bokstäver, siffror, _ och -')
+    .max(20, 'Produkt ID får inte vara längre än 20 tecken')
+    .optional()
+    .or(z.literal('')),
+  size: z
+    .string()
+    .regex(/^[A-Za-z0-9.,/\-xX]+$/, 'Storlek får endast innehålla bokstäver, siffror, . , / och -')
+    .max(10, 'Storlek får inte vara längre än 10 tecken')
+    .optional()
+    .or(z.literal('')),
+  productCount: z
+    .number({
+      required_error: 'Ange antal',
+      invalid_type_error: 'Ange ett giltigt antal',
+    })
+    .int('Antal måste vara ett heltal')
+    .min(1, 'Antal måste vara minst 1')
+    .max(10000, 'Antal får inte vara mer än 10 000'),
+  images: z
+    .array(z.custom<File>((file) => {
+      if (!file) return true // Optional field
+      if (!(file instanceof File)) return false
+      if (file.size > MAX_FILE_SIZE) return false
+      return ALLOWED_IMAGE_TYPES.includes(file.type)
+    }, {
+      message: 'Bilden måste vara mindre än 7MB och i formatet JPEG, PNG, WebP eller GIF',
+    }))
+    .max(MAX_IMAGE_COUNT, { message: `Maximalt ${MAX_IMAGE_COUNT} bilder kan bifogas` })
+    .optional()
+    .nullable(),
+  message: z
+    .string()
+    .max(2000, 'Meddelandet får inte vara längre än 2000 tecken')
+    .optional()
+    .or(z.literal('')),
+  gdprConsent: z
+    .boolean()
+    .refine(val => val === true, {
+      message: 'Du måste godkänna behandling av personuppgifter',
+    }),
+})
+
+/** TypeScript type from Zod schema */
+export type QuoteFormData = z.infer<typeof quoteFormSchema>
+
+// ===== TYPES =====
+
+/** Form field error type */
+export interface FormFieldError<T> {
+  field: keyof T
+  message: string
+}
+
+/** Form submission state */
+type FormState = 'idle' | 'submitting' | 'success' | 'error'
 
 // ===== COMPOSABLE =====
 /**
@@ -66,51 +157,40 @@ export function useQuoteForm() {
   // ===== METHODS =====
   /**
    * Validate a single field
-   * Dynamically imports validation logic to keep Zod out of critical bundle
    */
-  async function validateField(field: keyof QuoteFormData): Promise<boolean> {
+  function validateField(field: keyof QuoteFormData): boolean {
     try {
-      const { validateField: validate } = await import('~/utils/validation/quoteFormValidation')
-      const result = validate(field, formData.value)
-      if (result.valid) {
-        fieldErrors.value.delete(field)
-        return true
-      }
-      else {
-        if (result.error) {
-          fieldErrors.value.set(field, result.error)
-        }
-        return false
-      }
+      const fieldSchema = quoteFormSchema.shape[field]
+      fieldSchema.parse(formData.value[field])
+      fieldErrors.value.delete(field)
+      return true
     }
     catch (error) {
-      console.error('Field validation error:', error)
+      if (error instanceof z.ZodError) {
+        fieldErrors.value.set(field, error.errors[0].message)
+      }
       return false
     }
   }
 
   /**
    * Validate entire form
-   * Dynamically imports validation logic to keep Zod out of critical bundle
    */
-  async function validateForm(): Promise<boolean> {
+  function validateForm(): boolean {
     try {
-      const { validateForm: validate } = await import('~/utils/validation/quoteFormValidation')
-      const result = validate(formData.value)
-      
-      if (result.valid) {
-        fieldErrors.value.clear()
-        generalError.value = null
-        return true
-      }
-      else {
-        fieldErrors.value = new Map(result.errors)
-        return false
-      }
+      quoteFormSchema.parse(formData.value)
+      fieldErrors.value.clear()
+      generalError.value = null
+      return true
     }
     catch (error) {
-      console.error('Form validation error:', error)
-      generalError.value = 'Valideringsfel inträffade'
+      if (error instanceof z.ZodError) {
+        fieldErrors.value.clear()
+        error.errors.forEach((err) => {
+          const field = err.path[0] as keyof QuoteFormData
+          fieldErrors.value.set(field, err.message)
+        })
+      }
       return false
     }
   }
@@ -163,7 +243,7 @@ export function useQuoteForm() {
    * Submit form data to Netlify Forms
    */
   async function submitForm(): Promise<boolean> {
-    if (!await validateForm()) {
+    if (!validateForm()) {
       return false
     }
